BUG/MEDIUM: ingress: handle service port properly
in ingress object, check if user defined service port and
not target port. if he defined service port, find appropriate
target port.

+++ helper.go
@@ -67,7 +67,7 @@ func ConvertIngressRules(ingressRules []extensions.IngressRule) map[string]*Ingr
 				PathIndex:         pathIndex,
 				Path:              k8sPath.Path,
 				ServiceName:       k8sPath.Backend.ServiceName,
-				ServicePortInt:    k8sPath.Backend.ServicePort.IntValue(),
+				ServicePortInt:    int64(k8sPath.Backend.ServicePort.IntValue()),
 				ServicePortString: k8sPath.Backend.ServicePort.StrVal,
 				Status:            "",
 			}
